Improved: logic to hide login when confirming for better user experience
