+++ src/locales/en.json
@@ -12,5 +12,6 @@
   "Please fill in the OMS": "Please fill in the OMS",
   "Please fill in the user details": "Please fill in the user details",
   "Something went wrong while login. Please contact administrator.": "Something went wrong while login. Please contact administrator.",
+  "Sorry, your username or password is incorrect. Please try again.": "Sorry, your username or password is incorrect. Please try again.",
   "Username": "Username"
 }
